fix: (example and avm) post code review update
