+++ examples/p-chain/base.ts
@@ -1,29 +1,41 @@
-import { addTxSignatures } from '../../src/signer';
-import { TransferableOutput } from '../../src/serializable/avax';
-import { bech32ToBytes, hexToBuffer } from '../../src/utils';
-import { getContextFromURI } from '../../src/vms/context';
-import { newBaseTx } from '../../src/vms/pvm';
-import { pvmapi } from '../chain_apis';
+import { TransferableOutput, addTxSignatures, pvm, utils } from '../../src';
 import { getEnvVars } from '../utils/getEnvVars';
+import { setupEtnaExample } from './utils/etna-helper';
+
+/**
+ * The amount of AVAX to send to self.
+ */
+const SEND_AVAX_AMOUNT: number = 0.001;
 
 const main = async () => {
   const { AVAX_PUBLIC_URL, P_CHAIN_ADDRESS, PRIVATE_KEY } = getEnvVars();
 
-  const { utxos } = await pvmapi.getUTXOs({ addresses: [P_CHAIN_ADDRESS] });
-  const context = await getContextFromURI(AVAX_PUBLIC_URL);
+  const { context, feeState, pvmApi } = await setupEtnaExample(AVAX_PUBLIC_URL);
 
-  const tx = newBaseTx(context, [bech32ToBytes(P_CHAIN_ADDRESS)], utxos, [
-    TransferableOutput.fromNative(context.avaxAssetID, BigInt(0.1 * 1e9), [
-      bech32ToBytes(P_CHAIN_ADDRESS),
-    ]),
-  ]);
+  const { utxos } = await pvmApi.getUTXOs({ addresses: [P_CHAIN_ADDRESS] });
+
+  const tx = pvm.newBaseTx(
+    {
+      feeState,
+      fromAddressesBytes: [utils.bech32ToBytes(P_CHAIN_ADDRESS)],
+      outputs: [
+        TransferableOutput.fromNative(
+          context.avaxAssetID,
+          BigInt(SEND_AVAX_AMOUNT * 1e9),
+          [utils.bech32ToBytes(P_CHAIN_ADDRESS)],
+        ),
+      ],
+      utxos,
+    },
+    context,
+  );
 
   await addTxSignatures({
     unsignedTx: tx,
-    privateKeys: [hexToBuffer(PRIVATE_KEY)],
+    privateKeys: [utils.hexToBuffer(PRIVATE_KEY)],
   });
 
-  return pvmapi.issueSignedTx(tx.getSignedTx());
+  return pvmApi.issueSignedTx(tx.getSignedTx());
 };
 
 main().then(console.log);
