+++ examples/p-chain/delegate.ts
@@ -1,42 +1,49 @@
-import { PrimaryNetworkID } from '../../src/constants/networkIDs';
-import { addTxSignatures } from '../../src/signer';
-import { bech32ToBytes, hexToBuffer } from '../../src/utils';
-import { getContextFromURI } from '../../src/vms/context';
-import { PVMApi, newAddPermissionlessDelegatorTx } from '../../src/vms/pvm';
-import { pvmapi } from '../chain_apis';
+import { addTxSignatures, networkIDs, pvm, utils } from '../../src';
 import { getEnvVars } from '../utils/getEnvVars';
+import { setupEtnaExample } from './utils/etna-helper';
+
+const AMOUNT_TO_DELEGATE_AVAX: number = 1;
+const DAYS_TO_DELEGATE: number = 14;
 
 const main = async () => {
   const { AVAX_PUBLIC_URL, P_CHAIN_ADDRESS, PRIVATE_KEY } = getEnvVars();
 
-  const { utxos } = await pvmapi.getUTXOs({ addresses: [P_CHAIN_ADDRESS] });
-  const context = await getContextFromURI(AVAX_PUBLIC_URL);
-  const startTime = await new PVMApi().getTimestamp();
-  const startDate = new Date(startTime.timestamp);
-  const start = BigInt(startDate.getTime() / 1000);
-  const endTime = new Date(startTime.timestamp);
-  endTime.setDate(endTime.getDate() + 21);
-  const end = BigInt(endTime.getTime() / 1000);
-  const nodeID = 'NodeID-HKLp5269LH8DcrLvHPc2PHjGczBQD3td4';
+  const { context, feeState, pvmApi } = await setupEtnaExample(AVAX_PUBLIC_URL);
 
-  const tx = newAddPermissionlessDelegatorTx(
+  const { utxos } = await pvmApi.getUTXOs({ addresses: [P_CHAIN_ADDRESS] });
+
+  const startTime = await pvmApi.getTimestamp();
+  const startDate = new Date(startTime.timestamp);
+  const start: bigint = BigInt(startDate.getTime() / 1_000);
+
+  const endTime = new Date(startTime.timestamp);
+  endTime.setDate(endTime.getDate() + DAYS_TO_DELEGATE);
+  const end: bigint = BigInt(endTime.getTime() / 1_000);
+
+  // TODO: Get this from an argument.
+  const nodeId = 'NodeID-MqgFXT8JhorbEW2LpTDGePBBhv55SSp3M';
+
+  const tx = pvm.newAddPermissionlessDelegatorTx(
+    {
+      end,
+      feeState,
+      fromAddressesBytes: [utils.bech32ToBytes(P_CHAIN_ADDRESS)],
+      nodeId,
+      rewardAddresses: [utils.bech32ToBytes(P_CHAIN_ADDRESS)],
+      start,
+      subnetId: networkIDs.PrimaryNetworkID.toString(),
+      utxos,
+      weight: BigInt(AMOUNT_TO_DELEGATE_AVAX * 1e9),
+    },
     context,
-    utxos,
-    [bech32ToBytes(P_CHAIN_ADDRESS)],
-    nodeID,
-    PrimaryNetworkID.toString(),
-    start,
-    end,
-    BigInt(1e9),
-    [bech32ToBytes(P_CHAIN_ADDRESS)],
   );
 
   await addTxSignatures({
     unsignedTx: tx,
-    privateKeys: [hexToBuffer(PRIVATE_KEY)],
+    privateKeys: [utils.hexToBuffer(PRIVATE_KEY)],
   });
 
-  return pvmapi.issueSignedTx(tx.getSignedTx());
+  return pvmApi.issueSignedTx(tx.getSignedTx());
 };
 
 main().then(console.log);
