+++ examples/p-chain/export.ts
@@ -1,40 +1,42 @@
-import { TransferableOutput } from '../../src/serializable/avax';
-import { addTxSignatures } from '../../src/signer';
-import { bech32ToBytes, hexToBuffer } from '../../src/utils';
-import { getContextFromURI } from '../../src/vms/context';
-import { newExportTx } from '../../src/vms/pvm';
-import { pvmapi } from '../chain_apis';
-import { getChainIdFromContext } from '../utils/getChainIdFromContext';
+import { TransferableOutput, addTxSignatures, pvm, utils } from '../../src';
 import { getEnvVars } from '../utils/getEnvVars';
+import { setupEtnaExample } from './utils/etna-helper';
+
+const AMOUNT_TO_EXPORT_AVAX: number = 0.001;
 
 const main = async () => {
   const { AVAX_PUBLIC_URL, P_CHAIN_ADDRESS, PRIVATE_KEY, X_CHAIN_ADDRESS } =
     getEnvVars();
 
-  const context = await getContextFromURI(AVAX_PUBLIC_URL);
+  const { context, feeState, pvmApi } = await setupEtnaExample(AVAX_PUBLIC_URL);
 
-  const { utxos } = await pvmapi.getUTXOs({
+  const { utxos } = await pvmApi.getUTXOs({
     addresses: [P_CHAIN_ADDRESS],
   });
 
-  const tx = newExportTx(
+  const exportTx = pvm.newExportTx(
+    {
+      destinationChainId: context.xBlockchainID,
+      feeState,
+      fromAddressesBytes: [utils.bech32ToBytes(P_CHAIN_ADDRESS)],
+      outputs: [
+        TransferableOutput.fromNative(
+          context.avaxAssetID,
+          BigInt(AMOUNT_TO_EXPORT_AVAX * 1e9),
+          [utils.bech32ToBytes(X_CHAIN_ADDRESS)],
+        ),
+      ],
+      utxos,
+    },
     context,
-    getChainIdFromContext('X', context),
-    [bech32ToBytes(P_CHAIN_ADDRESS)],
-    utxos,
-    [
-      TransferableOutput.fromNative(context.avaxAssetID, BigInt(0.1 * 1e9), [
-        bech32ToBytes(X_CHAIN_ADDRESS),
-      ]),
-    ],
   );
 
   await addTxSignatures({
-    unsignedTx: tx,
-    privateKeys: [hexToBuffer(PRIVATE_KEY)],
+    unsignedTx: exportTx,
+    privateKeys: [utils.hexToBuffer(PRIVATE_KEY)],
   });
 
-  return pvmapi.issueSignedTx(tx.getSignedTx());
+  return pvmApi.issueSignedTx(exportTx.getSignedTx());
 };
 
 main().then(console.log);
