+++ src/vms/context/context.ts
@@ -33,8 +33,8 @@ export const getContextFromURI = async (
     pBlockchainID,
     cBlockchainID,
     avaxAssetID,
-    baseTxFee: BigInt(txFee),
-    createAssetTxFee: BigInt(createAssetTxFee),
+    baseTxFee: txFee,
+    createAssetTxFee: createAssetTxFee,
     networkID,
     hrp: getHRP(networkID),
     platformFeeConfig,
