+++ src/vms/pvm/index.ts
@@ -4,6 +4,6 @@ export * from './api';
 export * from './txs/fee';
 
 /**
- * @deprecated Exposed Etna builder functions under `e` namespace
+ * @deprecated PVM builder functions aliased under "e" are deprecated. Please use the builder functions on the root pvm export. Ex: pvm.e.newBaseTx -> pvm.newBaseTx.
  */
 export * as e from './etna-builder';
